+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 * Add support for shadow-cljs to `cider-jack-in`.
 * [#2244](https://github.com/clojure-emacs/cider/issues/2244): Display the REPL type in the modeline.
 * [#2238](https://github.com/clojure-emacs/cider/pull/2238): Allow specifying predicates for entries in `cider-jack-in-lein-plugins` and `cider-jack-in-nrepl-middlewares`.
-* Add support for test selectors. If test all or all loaded is called with a prefix ask for a selector in the minibuffer and only run those tests in the project which are marked with the selector. Add variation of test namespace which asks for a selector the same way and only runs a subset of the namespace tests.
+* Add support for test selectors. If test all or all loaded is called with a prefix ask for filter test selectors in the minibuffer and only run those tests in the project which match the filters. Add variation of test namespace which asks for filter selectors the same way and only runs a subset of the namespace tests.
 
 ### Bugs Fixed